+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/ThreadSafeEventChannelTests.m
@@ -7,61 +7,66 @@
 @import camera_avfoundation.Test;
 #endif
 @import XCTest;
-#import <OCMock/OCMock.h>
+
+#import "MockEventChannel.h"
 
 @interface ThreadSafeEventChannelTests : XCTestCase
+@property(readonly, nonatomic) MockEventChannel *mockEventChannel;
+@property(readonly, nonatomic) FLTThreadSafeEventChannel *threadSafeEventChannel;
 @end
 
 @implementation ThreadSafeEventChannelTests
 
-- (void)testSetStreamHandler_shouldStayOnMainThreadIfCalledFromMainThread {
-  FlutterEventChannel *mockEventChannel = OCMClassMock([FlutterEventChannel class]);
-  FLTThreadSafeEventChannel *threadSafeEventChannel =
-      [[FLTThreadSafeEventChannel alloc] initWithEventChannel:mockEventChannel];
+- (void)setUp {
+  [super setUp];
+  _mockEventChannel = [[MockEventChannel alloc] init];
+  _threadSafeEventChannel =
+      [[FLTThreadSafeEventChannel alloc] initWithEventChannel:_mockEventChannel];
+}
 
+- (void)testSetStreamHandler_shouldStayOnMainThreadIfCalledFromMainThread {
   XCTestExpectation *mainThreadExpectation =
       [self expectationWithDescription:@"setStreamHandler must be called on the main thread"];
   XCTestExpectation *mainThreadCompletionExpectation =
       [self expectationWithDescription:
                 @"setStreamHandler's completion block must be called on the main thread"];
-  OCMStub([mockEventChannel setStreamHandler:[OCMArg any]]).andDo(^(NSInvocation *invocation) {
+
+  [_mockEventChannel setSetStreamHandlerStub:^(NSObject<FlutterStreamHandler> *handler) {
     if (NSThread.isMainThread) {
       [mainThreadExpectation fulfill];
     }
-  });
+  }];
 
-  [threadSafeEventChannel setStreamHandler:nil
-                                completion:^{
-                                  if (NSThread.isMainThread) {
-                                    [mainThreadCompletionExpectation fulfill];
-                                  }
-                                }];
+  [_threadSafeEventChannel setStreamHandler:nil
+                                 completion:^{
+                                   if (NSThread.isMainThread) {
+                                     [mainThreadCompletionExpectation fulfill];
+                                   }
+                                 }];
   [self waitForExpectationsWithTimeout:1 handler:nil];
 }
 
 - (void)testSetStreamHandler_shouldDispatchToMainThreadIfCalledFromBackgroundThread {
-  FlutterEventChannel *mockEventChannel = OCMClassMock([FlutterEventChannel class]);
-  FLTThreadSafeEventChannel *threadSafeEventChannel =
-      [[FLTThreadSafeEventChannel alloc] initWithEventChannel:mockEventChannel];
-
   XCTestExpectation *mainThreadExpectation =
       [self expectationWithDescription:@"setStreamHandler must be called on the main thread"];
   XCTestExpectation *mainThreadCompletionExpectation =
       [self expectationWithDescription:
                 @"setStreamHandler's completion block must be called on the main thread"];
-  OCMStub([mockEventChannel setStreamHandler:[OCMArg any]]).andDo(^(NSInvocation *invocation) {
+
+  [_mockEventChannel setSetStreamHandlerStub:^(NSObject<FlutterStreamHandler> *handler) {
     if (NSThread.isMainThread) {
       [mainThreadExpectation fulfill];
     }
-  });
+  }];
 
+  __weak typeof(self) weakSelf = self;
   dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
-    [threadSafeEventChannel setStreamHandler:nil
-                                  completion:^{
-                                    if (NSThread.isMainThread) {
-                                      [mainThreadCompletionExpectation fulfill];
-                                    }
-                                  }];
+    [weakSelf.threadSafeEventChannel setStreamHandler:nil
+                                           completion:^{
+                                             if (NSThread.isMainThread) {
+                                               [mainThreadCompletionExpectation fulfill];
+                                             }
+                                           }];
   });
   [self waitForExpectationsWithTimeout:1 handler:nil];
 }
@@ -69,11 +74,13 @@
 - (void)testEventChannel_shouldBeKeptAliveWhenDispatchingBackToMainThread {
   XCTestExpectation *expectation =
       [self expectationWithDescription:@"Completion should be called."];
-  dispatch_async(dispatch_queue_create("test", NULL), ^{
-    FLTThreadSafeEventChannel *channel = [[FLTThreadSafeEventChannel alloc]
-        initWithEventChannel:OCMClassMock([FlutterEventChannel class])];
 
-    [channel setStreamHandler:OCMOCK_ANY
+  __weak typeof(self) weakSelf = self;
+  dispatch_async(dispatch_queue_create("test", NULL), ^{
+    FLTThreadSafeEventChannel *channel =
+        [[FLTThreadSafeEventChannel alloc] initWithEventChannel:weakSelf.mockEventChannel];
+
+    [channel setStreamHandler:nil
                    completion:^{
                      [expectation fulfill];
                    }];