+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation/include/camera_avfoundation/CameraPlugin_Test.h
@@ -6,7 +6,14 @@
 
 #import "CameraPlugin.h"
 #import "FLTCam.h"
+#import "FLTCamConfiguration.h"
+#import "FLTCameraDeviceDiscovering.h"
+#import "FLTCaptureDevice.h"
 #import "messages.g.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+typedef id<FLTCaptureDevice> _Nonnull (^CaptureNamedDeviceFactory)(NSString *name);
 
 /// APIs exposed for unit testing.
 @interface CameraPlugin ()
@@ -15,20 +22,25 @@
 @property(nonatomic, strong) dispatch_queue_t captureSessionQueue;
 
 /// An internal camera object that manages camera's state and performs camera operations.
-@property(nonatomic, strong) FLTCam *camera;
+@property(nonatomic, strong) FLTCam *_Nullable camera;
 
 /// Inject @p FlutterTextureRegistry and @p FlutterBinaryMessenger for unit testing.
-- (instancetype)initWithRegistry:(NSObject<FlutterTextureRegistry> *)registry
-                       messenger:(NSObject<FlutterBinaryMessenger> *)messenger;
+- (nonnull instancetype)initWithRegistry:(NSObject<FlutterTextureRegistry> *_Nullable)registry
+                               messenger:(NSObject<FlutterBinaryMessenger> *_Nullable)messenger;
 
 /// Inject @p FlutterTextureRegistry, @p FlutterBinaryMessenger, and Pigeon callback handler for
 /// unit testing.
-- (instancetype)initWithRegistry:(NSObject<FlutterTextureRegistry> *)registry
-                       messenger:(NSObject<FlutterBinaryMessenger> *)messenger
-                       globalAPI:(FCPCameraGlobalEventApi *)globalAPI NS_DESIGNATED_INITIALIZER;
+- (nonnull instancetype)initWithRegistry:(NSObject<FlutterTextureRegistry> *_Nullable)registry
+                               messenger:(NSObject<FlutterBinaryMessenger> *_Nullable)messenger
+                               globalAPI:(FCPCameraGlobalEventApi *_Nullable)globalAPI
+                        deviceDiscoverer:(id<FLTCameraDeviceDiscovering>)deviceDiscoverer
+                           deviceFactory:(CaptureNamedDeviceFactory)deviceFactory
+                   captureSessionFactory:(CaptureSessionFactory)captureSessionFactory
+               captureDeviceInputFactory:(id<FLTCaptureDeviceInputFactory>)captureDeviceInputFactory
+    NS_DESIGNATED_INITIALIZER;
 
 /// Hide the default public constructor.
-- (instancetype)init NS_UNAVAILABLE;
+- (nonnull instancetype)init NS_UNAVAILABLE;
 
 /// Called by the @c NSNotificationManager each time the device's orientation is changed.
 ///
@@ -40,7 +52,10 @@
 /// @param name the name of the camera.
 /// @param settings the creation settings.
 /// @param completion the callback to inform the Dart side of the plugin of creation.
-- (void)createCameraOnSessionQueueWithName:(NSString *)name
+- (void)createCameraOnSessionQueueWithName:(NSString *_Nullable)name
                                   settings:(FCPPlatformMediaSettings *)settings
-                                completion:(void (^)(NSNumber *, FlutterError *))completion;
+                                completion:(void (^)(NSNumber *_Nullable,
+                                                     FlutterError *_Nullable))completion;
 @end
+
+NS_ASSUME_NONNULL_END