+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation/include/camera_avfoundation/FLTCaptureDevice.h
@@ -1,0 +1,87 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+@import AVFoundation;
+@import Foundation;
+
+NS_ASSUME_NONNULL_BEGIN
+
+@protocol FLTCaptureDevice <NSObject>
+
+// Device
+- (NSString *)uniqueID;
+
+// Position/Orientation
+- (AVCaptureDevicePosition)position;
+
+// Format/Configuration
+- (AVCaptureDeviceFormat *)activeFormat;
+- (NSArray<AVCaptureDeviceFormat *> *)formats;
+- (void)setActiveFormat:(AVCaptureDeviceFormat *)format;
+
+// Flash/Torch
+- (BOOL)hasFlash;
+- (BOOL)hasTorch;
+- (BOOL)isTorchAvailable;
+- (AVCaptureTorchMode)torchMode;
+- (void)setTorchMode:(AVCaptureTorchMode)torchMode;
+- (BOOL)isFlashModeSupported:(AVCaptureFlashMode)mode;
+
+// Focus
+- (BOOL)isFocusPointOfInterestSupported;
+- (BOOL)isFocusModeSupported:(AVCaptureFocusMode)mode;
+- (void)setFocusMode:(AVCaptureFocusMode)focusMode;
+- (void)setFocusPointOfInterest:(CGPoint)point;
+
+// Exposure
+- (BOOL)isExposurePointOfInterestSupported;
+- (void)setExposureMode:(AVCaptureExposureMode)exposureMode;
+- (void)setExposurePointOfInterest:(CGPoint)point;
+- (float)minExposureTargetBias;
+- (float)maxExposureTargetBias;
+- (void)setExposureTargetBias:(float)bias completionHandler:(void (^_Nullable)(CMTime))handler;
+- (BOOL)isExposureModeSupported:(AVCaptureExposureMode)mode;
+
+// Zoom
+- (float)maxAvailableVideoZoomFactor;
+- (float)minAvailableVideoZoomFactor;
+- (float)videoZoomFactor;
+- (void)setVideoZoomFactor:(float)factor;
+
+// Camera Properties
+- (float)lensAperture;
+- (CMTime)exposureDuration;
+- (float)ISO;
+
+// Configuration Lock
+- (BOOL)lockForConfiguration:(NSError **)error;
+- (void)unlockForConfiguration;
+
+// Frame Duration
+- (CMTime)activeVideoMinFrameDuration;
+- (void)setActiveVideoMinFrameDuration:(CMTime)duration;
+- (CMTime)activeVideoMaxFrameDuration;
+- (void)setActiveVideoMaxFrameDuration:(CMTime)duration;
+
+@end
+
+@protocol FLTCaptureInput <NSObject>
+@property(nonatomic, readonly) NSArray<AVCaptureInputPort *> *ports;
+@end
+
+@protocol FLTCaptureDeviceInputFactory <NSObject>
+- (nullable id<FLTCaptureInput>)deviceInputWithDevice:(id<FLTCaptureDevice>)device
+                                                error:(NSError **)error;
+@end
+
+@interface AVCaptureDevice (FLTCaptureDevice) <FLTCaptureDevice>
+@end
+
+@interface AVCaptureInput (FLTCaptureInput) <FLTCaptureInput>
+@end
+
+@interface FLTDefaultCaptureDeviceInputFactory : NSObject <FLTCaptureDeviceInputFactory>
+@end
+
+NS_ASSUME_NONNULL_END