+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.9.18+2
+
+* Introduces new protocols `FLTEventChannel` and `FLTCameraDeviceDiscovering`
+
+## 0.9.18+1
+
+* Introduces new protocols `FLTCaptureDeviceControlling` and `FLTDeviceOrientationProviding`
+
 ## 0.9.18
 
 * Adds API support query for image streaming.